Obtain the GTFS files
