+++ lib/gtfs.rb
@@ -1,0 +1,5 @@
+require_relative 'gtfs/setup'
+
+module GTFS
+  include Setup
+end
